Error out if dependent flags are provided without their parent (#6780)
We have a few flags that only make sense in the context of others.
People are mistakenly using these flags on their own and correctly
observing that nothing happens. This change records those deps such that
people using invalid combinations of flags are warned.

This also deprecates some old flags, and reorganises the args such that
they are closer to one-another in the help doc.