fix: avoid retry panic (#6926)
### Description

As reported in #6921 `turbo` is panicking when making HTTP requests.
This PR doesn't fix the user issue, but it does avoid the panic and
provide a better warning message to the end user.

The panic can be caused in 2 ways:
- Trying to retry a request that has a streaming body. This wasn't
happening, but I changed the code so we at least don't panic if we
accidentally call this function with a streaming body.
- Constructing a `RequestBuilder` with an invalid url (see
https://github.com/seanmonstar/reqwest/issues/1943) will cause the
`try_clone` method to return `None` causing us to panic. This can easily
be done if a user passes an invalid API url. We mitigate this by now
only constructing `RequestBuilder` with `Url` which has already been
verified instead of `&str` which delays the verification and obscures
the failure.

Minor change is removing the `do_preflight` method from the client trait
as it is an implementation detail and it reduced the amount of places
that needed to get updated.

### Testing Instructions

Create a bogus entry in `auth.json` e.g. 
```
[0 olszewski@chriss-mbp] /Users/olszewski/code/vercel/turborepo $ bat ~/Library/Application\ Support/turborepo/auth.json 
───────┬─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
       │ File: /Users/olszewski/Library/Application Support/turborepo/auth.json
───────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1   │ {
   2   │   "tokens": {
   3   │     "junk-url": "junk-token"
   4   │   }
   5   │ }
───────┴─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

On main verify that the current experience is far from ideal:
<img width="956" alt="Screenshot 2024-01-05 at 1 26 09 PM"
src="https://github.com/vercel/turbo/assets/4131117/f63c60f1-db0f-441f-a386-11e4078cd55c">

Build using these changes and verify that there's a better UX:

<img width="1038" alt="Screenshot 2024-01-05 at 1 30 05 PM"
src="https://github.com/vercel/turbo/assets/4131117/13fb3d21-bf48-44b9-9586-089d03afa8ee">



Closes TURBO-2002