chore: add reason for service being unavailable (#6993)
### Description

Having trouble making progress on #6934 and the daemon logs aren't
especially helpful. Tonic is saying the service is unavailable, but I'm
unsure of what would cause it to become unavailable. From the docs:

> The service is currently unavailable. This is a most likely a
transient condition and may be corrected by retrying with a back-off.

From our server code, I'm not seeing us explicitly throw this anywhere.

This PR does a few things:
- Change the warning to use the default formatter for the error instead
of the debug formatter
- Plumb the message from the `tonic::Status` event into our
`Unavailable` error variant

### Testing Instructions

👀 

Closes TURBO-2034