fix(prune): correctly create symlinks to directories  (#6983)
### Description

Fixes #5842

> The behavior of RecursiveCopy was written prior to cache-to-tar which
removed all other consumers of this functionality.

This is why there was the odd behavior with copying symlink'd
directories. We now only use this function for `prune` so making this
behavior change makes sense instead of keeping the quirk.

Note:
This does not handle making sure the destination is copied over as well.
This could be implemented, but it isn't a simple fix. An easy workaround
it to make sure the files/directories are part of a package and make
sure that package is declared as a dependency.

### Testing Instructions

Updated unit test to verify new behavior

Closes TURBO-2027

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>