Use `exports`  in Internal Packages doc. (#7079)
### Description

While `main` and `types` works, it often corners our users into using
barrel files. Instead, let's start our users off on a path that allows
them to create many entrypoints to their packages if their repo were to
grow to need it).

NOTE: We're pushing past broken CI tests here. We're aware of the breakage and working to fix it.