fix: filter package.json files that are subfolders of other packages (#7025)
### Description

We have an issue in the rust port with an overeager glob walking
implementation. We want to stop walking when a package.json is
encountered such that sub-packages are ignored. This is not possible
unfortunately due to the traversal order in `walkdir`, so instead we
must filter the values out afterwards.

### Testing Instructions

New integration test.

Co-authored-by: Alexander Lyon <Alexander Lyon>