Turbopack: use inline sourcemaps for HMR chunks
Previously, we sent a URL with a query string identifying the module, but this never worked in nextturbo. Instead, this sends the source map string itself, where it is then encoded as base64 and used inline in the `eval`ed code.

Test Plan: Updated a client component and verified browser devtools shows updated source code through the source map.