Print telemetry warning on stderr
This fixes an issue with unparsable output from `--dry=json` that
causes `turbo-ignore` to fail with a JSON parse error if the user has
not disabled the message.

To reproduce the issue this PR seeks to resolve run the following
commands in any turbo enabled repo.

```
rm ~/.config/turborepo/telemetry.json
pnpx turbo-ignore # fails with JSON parse error
```