chore(turborepo): Ensure fresh builds (#7321)
### Description

This had been bugging me for a while. When building turbo, even with no
changes, we'd still rebuild turborepo-lib and turbo. Turns out this is
cause we were generating `file_descriptor_set.bin` into the same
directory we were watching for protobuf changes. To avoid this, I put
the protobuf files in a nested directory.

### Testing Instructions

Compile and be amazed at the speed!


Closes TURBO-2294