feat(tasks-fs): bump up notify, fix empty file watch (#7340)
### Description

This PR upgrades `notify` pkg to create a filesystem watcher. 

It is due to recent issue we found in PACK-2437, the edge cases like

- start a watcher with file have any contents
- open editor, clear all (make file empty), save

doesn't create any file watcher event. This makes some cases turbopack
does not trigger hmr even if it's desired.

One thing to note is even with upgrade its event kind is somewhat
unexpected; it's not `Modify` event but `Metadata` event. I can't say
why it emits in that way, but add it as workaround for now.


We had experiences of trying to upgrade & revert this pkg before, so the
upgrade plan need some caution.

Plan is

- Once PR is approved, cut turbopack _before_ PR and bump next.js first
- Land PR, create new turbopack release so release contains only 1
changes to notify
- Ensure new turbopack update in next.js won't break (like upgrading
front, etcs) and make it easy to revert


Closes PACK-2437.