feat(@turbo/repository): add affectedPackages API (#7326)
This API takes a list of files (strings anchored to the root of the
workspace) and returns a set of packages that were affected by that
change. Importantly, this does _not_ include the whole graph of
packages affected by these files changes, so it's a pretty simple
mapping of files to the packages they belong to. If a known global file
changes, all packages are returned.

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>