fix(@turbo/repository): run JS native package tests when turbo core changes (#7374)
This changes CI setup to run JS native packages when turbo core
changes. There is only one JS native package. This also

- excludes this package from the other JS package tests in CI
- ensures that turbo core is built before running these tests

This new workflow does not handle skipping these tests when
relevant files have not changes. This does mean it will block unrelated
PRs, but we can resolve that if it's a problem.