Audit daemon error paths and ensure we are doing everything we can to handle (#7334)
### Description

We have a few integration points with the daemon, this PR ensures they
are all nicely handled.

- [x] `cache.rs`: we emit a debug warning and a telemetry event in the
case we are unable to mark outputs correctly
- [x] `lsp/lib.rs`: in the LSP, package discovery means we cannot
fulfill the user's request, so we must return an internal error. the
error message has been set. Additionally, the LSP is a case where
waiting for an answer is actually desirable, so a new API to support
this has been added.
- [x] package discovery: if we have a bug in watching, we may yield
packages that don't exist. we need to report a telemetry event, in this
case so that we can be aware of when this happens
- [x] daemon interaction in run codepath: when trying to use the daemon
during run we want to debug log but not error
- [x] daemon connection in daemon codepath: when calling the daemon we
want errors to be explicitly reported. in this case, the errors bubbles
all the way up and is printed out as expected

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->


Closes TURBO-2319

---------

Co-authored-by: Alexander Lyon <Alexander Lyon>