fix(turbopack): Fix `pure` lint for CSS Modules in lightningcss mode (#7431)
### Description

Fix `pure` lint in `lightningcss` mode. `lightningcss` does not provide
location information of selectors, so we cannot provide source location.
That's limitation of `cssparser` and `selectors` from `servo` project.
It was the reason I didn't use them while building `swc_css`. As a last
resort, we can use `Debug` implementation to print the selectors.


### Testing Instructions

Trying it on the internal apps is the best way.

Closes PACK-2546