chore: Change daemon log location to be in repo (#7468)
### Description

We were previously logging to a file outside the repo. This made it
rather annoying to find, especially if the daemon is not running. Now we
log to the `.turbo/daemon` folder inside the repository. Also, on `turbo
daemon logs`, if the daemon isn't running, we fall back to looking for
the latest file in `.turbo/daemon`.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->


Closes TURBO-2417