feat: allow for faux comments at top level of turbo.json (#7546)
### Description

Closes #7522

All this does is skip us throwing if we encounter a top level `//` key
in the `turbo.json`. I personally thinks this makes some sense since the
`//` key is used across the JS ecosystem for comments in JSON files. We
do support JSONC for `turbo.json`, but that might not be immediately
obvious without reading documentation.

### Testing Instructions

Added unit test


Closes TURBO-2462