fix: support pnpm workspace path aliasing (#7512)
### Description

Fixes #6744 

I highly suggest reviewing each commit on it's own as the code move done
in the first commit makes the cumulative diff hard to understand. I
wanted to clean up this code a bit as we'll need to extend it further
for supporting https://github.com/vercel/turbo/issues/7188

This adds support for the PNPM feature of [aliasing a package via
path](https://pnpm.io/workspaces#referencing-workspace-packages-through-their-relative-path).

The slightly tricky part of this is that there are many valid specifiers
for the `workspace:` protocol. We currently support `workspace:1.2.3`,
`workspace:*`,
[`workspace:bar@*`](https://pnpm.io/workspaces#referencing-workspace-packages-through-aliases),
`workspace:../path/to/package`.

### Testing Instructions

Added unit tests for the case of a dependency of the form `"foo":
"workspace:../bar"` where `foo` is another workspace package.

Verified fix works with repro provided in:
https://github.com/vercel/turbo/issues/6744