fix: vercel auth file borks login (#7556)
### Description
Because `vc logout` doesn't remove it's pseudo-comments, deserialize breaks when trying to extract the token.
This PR fixes that by setting the deserialized struct to have an optional token field and then we make sure the extraction didn't encounter an error _and_ that the token is `Some`.

### Testing Instructions
First, make sure you build the branch.
1. `vc login && vc logout` -> make the Vercel auth file and then remove the `token` field. You could do this manually as well.
2. `turbo login` -> Shouldn't error and ask for a new login.


Closes TURBO-2485