Remove barrel file from design-system-example (#7580)
### Description

As @lorcan-codes pointed out in #7529, the `design-system` example could
quickly run into performance issues. This was because we were using a
barrel file.

I've swapped the `main` field `./packages/ui/package.json` for an
`exports` field that allows for creating individual entrypoints for the
package to remove this class of issue.

### Testing Instructions

I verified that `turbo build && cd apps/docs && pnpm preview-storybook`
runs correctly. Additionally, I made sure that `turbo dev` works and
that edits to both the `Button` component itself and the story for it
show hot-reloaded changes.

Happy to have someone else check me!


Closes TURBO-2505