refactor(turborepo): Package Detection (#7549)
### Description
Factored out package detection, i.e. the part where we take a file and
map it to the package into a trait. This allows us to have multiple
implementations, such as a default one that only uses the package graph,
and a more advanced implementation that uses the package graph and
turbo.json.

<!--
  ✍️ Write a short summary of your work.
  If necessary, include relevant screenshots.
-->

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->


Closes TURBO-2465

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>