[LSP] Attempt to detect local turbo path from package manager as fallback (#7662)
As identified in #7425, we can't always assume the people are able to install turbo globally, meaning we can't rely on turbo's build-in resolution in call cases. As a fallback, this will attempt to resolve the path from your package manager. It will _still_ surface a prompt to install global turbo however, unless you silence the recommendation with the "Use Local Turbo" setting. 

Also bundled in with this change is an attempt to support asdf as well by sourcing its script before finding the turbo executable.

Closes TURBO-2564