ci(workflow): expands more wasi targets (#7862)
### TL;DR

This pull request updates the build flags for the 'turbopack' wasm32-wasi target and increases the number of target packages.

### What changed?

In the GitHub workflow file for tests, the `cargo check` command for the 'turbopack-wasi' target now specifies the `CARGO_BUILD_TARGET` environment variable. Additionally, in the 'Cargo.toml' file, the list of packages that can be compiled against the wasm32-wasi target has been expanded to include more turbo-task related crates.

### How to test?

To test these changes, run the GitHub workflows and check whether the 'turbopack' build process and the associated tests succeed.

### Why make this change?

Updating the build flags and target packages for the wasm32-wasi build process in 'turbopack' may lead to better performance and broader feature support.