Docs(deploying-with-docker.mdx): Inform user about `turbo prune` command
The command `turbo prune web --docker` produces an error if the name in the package.json is not set to web

```
{
   name: "nextjs-app"
}
```

It will then throw the error: `package web not found`. This change informs the user that it is being assumed that the name given in the package.json is "web".