feat: Re-run dependent tasks in watch mode (#7898)
Instead of using `--filter` in our re-running of tasks for watch mode,
we instead prune the task graph to run the changed task and its
dependent tasks. This has the benefit of following the task graph and
not the package graph.

You can verify this works by creating a task dependency relation that is
not in the package graph. So, make a `ui#test` task that depends on a
`docs#test` task, but don't have `ui` depend on `docs`. Then verify that
when the `docs` are changed, we re-run the `ui#test` task too.

Closes TURBO-2752