add some more logging to the extension startup (#8047)
### Description

Adds a logging channel for the turbo vsc extension and rework some logging to prevent noisiness in the notification bar. This also adds some extra handling for homebrew based npm installs, and some additional handling to ensure paths are trimmed properly.

### Testing Instructions

Build and boot up the extension and observe the new output tab for Turbo.

Closes TURBO-2894